+++ samples/HelloWorld.Editor.Demo/DemoPage.cs
@@ -1,0 +1,17 @@
+using Avalonia.Controls;
+using JLeb.Estragonia;
+
+namespace HelloWorld.Editor.Demo;
+
+public sealed class DemoPage : IEditorPage {
+
+	public string Id
+		=> "estragonia.demo";
+
+	public string Title
+		=> "Estragonia";
+
+	public Control Create()
+		=> new DemoView();
+
+}